Move a function to a file where it makes more sense
