+++ src/loadsave_gui.cpp
@@ -11,110 +11,6 @@
 #include "window.h"
 #include "gamecontrol.h"
 
-/* Includes and definitions needed by ListDirectory. This can probably be shortened, but it would need testing on all platforms. */
-#include <cerrno>
-#ifdef _WIN32
-#ifdef _MSC_VER
-#include <cstdio>
-#else
-#include <cstdint>
-#endif
-#endif
-#include <cstdlib>
-#include <cstring>
-#include <sys/stat.h>
-#ifdef _WIN32
-#include <dos.h>
-#include <windows.h>
-#ifdef _MSC_VER
-#include <corecrt_io.h>
-#define S_ISDIR(x) ((x & _S_IFDIR) ? 1 : 0)
-#endif
-#else  // not _WIN32
-#include <fcntl.h>
-#include <glob.h>
-#include <sys/mman.h>
-#include <sys/statvfs.h>
-#include <sys/types.h>
-#include <unistd.h>
-#endif
-
-/**
- * List all files in a directory.
- * Code copied from Widelands (https://github.com/widelands/widelands/tree/master/src/io/filesystem/disk_filesystem.cc) with gratitude
- * (with some minor changes to make it compile here).
- * @param path Directory path.
- * @return All files in the directory.
- */
-static std::set<std::string> ListDirectory(const std::string& path) {
-	const std::string directory_ = ".";
-	const std::string root_ = directory_;
-
-#ifdef _WIN32
-	std::string buf;
-	struct _finddata_t c_file;
-	intptr_t hFile;
-
-	if (path.size())
-		buf = directory_ + '\\' + path + "\\*";
-	else
-		buf = directory_ + "\\*";
-
-	std::set<std::string> results;
-
-	hFile = _findfirst(buf.c_str(), &c_file);
-	if (hFile == -1)
-		return results;
-
-	std::string realpath = path;
-
-	if (!realpath.empty()) {
-		realpath.append("\\");
-	}
-	do {
-		if ((strcmp(c_file.name, ".") == 0) || (strcmp(c_file.name, "..") == 0)) {
-			continue;
-		}
-		const std::string filename = realpath + c_file.name;
-		std::string result = filename.substr(root_.size() + 1);
-
-		// Paths should not contain any windows line separators.
-		for (size_t i = 0; i < result.size(); i++) if (result.at(i) == '\\') result.at(i) = '/';
-		results.insert(result);
-	} while (_findnext(hFile, &c_file) == 0);
-
-	_findclose(hFile);
-
-	return results;
-#else
-	std::string buf;
-	glob_t gl;
-	int32_t ofs;
-
-	if (!path.empty()) {
-		buf = directory_ + '/' + path + "/*";
-		ofs = directory_.length() + 1;
-	} else {
-		buf = directory_ + "/*";
-		ofs = directory_.length() + 1;
-	}
-	std::set<std::string> results;
-
-	if (glob(buf.c_str(), 0, nullptr, &gl)) {
-		return results;
-	}
-
-	for (size_t i = 0; i < gl.gl_pathc; ++i) {
-		const std::string filename(&gl.gl_pathv[i][ofs]);
-		results.insert(filename.substr(root_.size() + 1));
-	}
-
-	globfree(&gl);
-
-	return results;
-#endif
-}
-
 /**
  * Game loading/saving gui.
  * @ingroup gui_group
